Make general improvements to documentation (#15)
