+++ docs/index.html
@@ -1,0 +1,6 @@
+<!DOCTYPE html>
+<html>
+    <head>
+        <meta http-equiv="refresh" content="0; url=build/html/index.html" />
+    </head>
+</html>
