+++ translations/views.py
@@ -1,5 +1,5 @@
 """
-The views for the `translations`app.
+The views for the translations app.
 
 This will be the only interface that consumers of this service will
 communicate with.
@@ -8,29 +8,11 @@ The purpose of this will be to act as a bridge between the decyphr application
 and the Google Translate API. Incoming text will be deserialized, translated
 and returned back the cosumer that will ultimately be returned to an end user.
 """
-from django.conf import settings
 from rest_framework import viewsets
 from rest_framework import status
 from rest_framework.response import Response
-from google.cloud import translate
+from .google_wrapper import GoogleMixin
 from .serializers import PlainTextTranslationSerializer
-
-
-class GoogleMixin:
-
-    translation_client = translate.TranslationServiceClient()
-    parent = translation_client.location_path(
-        settings.GOOGLE_PROJECT, "global")
-
-    def translate_text(self, target_language_code, text):
-        response = self.translation_client.translate_text(
-            parent=self.parent,
-            contents=[text],
-            mime_type="text/plain",
-            target_language_code=target_language_code
-        )
-
-        return response.translations[0].translated_text
 
 
 class PlainTextTranslator(viewsets.ViewSet, GoogleMixin):
@@ -48,17 +30,20 @@ class PlainTextTranslator(viewsets.ViewSet, GoogleMixin):
         Google and it will be return the client
 
         Args:
-            language_code (str): A ISO 639-1 code that will be used to inform
+            language_code (str): A ISO 639-1 code that will be used to inform \
             Google Translate of the target language
             text (str): The text that is needed to be translated
         
         Returns:
             JSON:
-                language_code (str): A ISO 639-1 code that will be used to
-                inform Google Translate of the target language
+                language_code (str): A ISO 639-1 code of the target language
                 text (str): The text that is needed to be translated
-                translated_text (str): The text that has been translated into
-                the targetted language
+                translated_text (str): The translated text
+        
+        Example:
+            curl -X POST -H 'Content-type: application/json' \
+                http://127.0.0.1:8000/api/v1/plain-text/ \
+                -d '{"language_code": "en", "text": "oi"}'
         """
         language_code = request.data["language_code"]
         text = request.data["text"]
